Use a per-version cache file for the index state
cabal-install will now use a version suffixed cache file for the index
state.

If you are regularly changing between cabal-install versions, this will
be less annoying as you won't have to regenerate the cache each time you
switch project.

There is one tricky part of the implementation. If you update the index
with a newer cabal-install, then the old-style cabal-install caches are
invalidated by replacing them with an empty file.
This is because in cabal-install (until this commit), the freshness of
the cache was now checked by `readIndexCache`.

If you update with an older `cabal-install` then the freshness check
will see the cache for your new cabal-install is older than the index,
and update it.

Fixes #7502