Fix -haddock swallowing other GHC arguments
When 'documentation: true' was enabled, the code was adding '-haddock' to the front
of programOverrideArgs list. Due to how MapMappend combines argument lists using
the <> operator (which is left-biased), this caused the -haddock argument to
override all other arguments.

Fixed by appending -haddock to the end of programOverrideArgs instead of prepending
it, ensuring it doesn't interfere with other arguments.

Added a test script to verify both -haddock and other arguments (like -O2)
are properly preserved.